backend/requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
pymongo
python-dotenv
bcrypt
pyjwt
pydantic
passlib
certifi
pytest
httpx
requests
websockets